requirements.txt
fastapi
openai-whisper
chattts
torch
torchaudio
webrtcvad
aiortc
uvicorn